# book-store/orders/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service for placing and reading book orders",
        "title": "Book Store Orders Service",
        "contact": {
            "name": "Dimas",
            "url": "https://github.com/Vesninovich",
            "email": "dmitry@vesnin.work"
        },
        "license": {
            "name": "ISC",
            "url": "https://www.isc.org/licenses/"
        },
        "version": "0.0"
    },
    "host": "localhost:8004",
    "basePath": "/",
    "paths": {
        "/order": {
            "post": {
                "description": "place new book order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "place order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "order data",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github.com_Vesninovich_go-tasks_book-store_orders_rest.createAPIModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "created order",
                        "schema": {
                            "$ref": "#/definitions/github.com_Vesninovich_go-tasks_book-store_orders_rest.apiModel"
                        }
                    },
                    "400": {
                        "description": "malformed book id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "requested book not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/order/{id}": {
            "get": {
                "description": "get order by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "get order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "requested order",
                        "schema": {
                            "$ref": "#/definitions/github.com_Vesninovich_go-tasks_book-store_orders_rest.apiModel"
                        }
                    },
                    "400": {
                        "description": "malformed id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "requested order not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "change order description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "update description",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "new description",
                        "name": "description",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github.com_Vesninovich_go-tasks_book-store_orders_rest.descUpdAPIModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "updated order",
                        "schema": {
                            "$ref": "#/definitions/github.com_Vesninovich_go-tasks_book-store_orders_rest.apiModel"
                        }
                    },
                    "400": {
                        "description": "malformed order id or bad data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "requested order not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "remove order",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "remove order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order id",
                        "name": "order",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "removed order",
                        "schema": {
                            "$ref": "#/definitions/github.com_Vesninovich_go-tasks_book-store_orders_rest.apiModel"
                        }
                    },
                    "400": {
                        "description": "malformed order id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "requested order not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github.com_Vesninovich_go-tasks_book-store_orders_rest.apiModel": {
            "type": "object",
            "properties": {
                "bookID": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "github.com_Vesninovich_go-tasks_book-store_orders_rest.createAPIModel": {
            "type": "object",
            "properties": {
                "bookID": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        },
        "github.com_Vesninovich_go-tasks_book-store_orders_rest.descUpdAPIModel": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                }
            }
        }
    },
    "tags": [
        {
            "description": "Requesting and placing orders",
            "name": "Order"
        }
    ]
}